Web Notes App/backend/requirements.txt
Django
python-dotenv
djangorestframework
django-cors-headers
